# perf(python): Directly access the key-value pairs of a dict